Various bug fixes related to endianness, etc.
The main one is that tri files read from b4 or lb4 (so, all of them) were
stored internally in Python in the same format.  That's a problem at write
time because the C versions don't know to convert them.  For now I'm just
performing conversions to system defaults at read time