Break out axes extent finders
The main purpose of this is to make one function that finds the extents
of a figure so that another one can properly deconflict with it.  This will
finally create an interface to do the seam plots below a line load plot that
is not full of gremlins.  That interface has not actually been written yet.

Also, using pixels for margins might be required if I ever plan to use
SubplotCols and SubplotCols; scaling might not give enough room for labels
the way it's currently written.