Update main worksheet header reader to interpret array doctrine
This doesn't add any full functionality because read_xls_firstrowtypes()
doesn't support arrays yet.  But debugging showed that the header reader is
working properly.  Integrating read_xls_header() with read_xls_first...()
will also expand the doctrine to work with missing array lengths by
assuming an array is implied when the header has missing columns *if* the
first data row has the same type.

This function really needs to be split into parts to be more in line with
the 'Zen of Python' principle 'Complex is better than complicated'