Do most of plot_mpl -> mpl breakout
This is a much more workable plot utilities architecture.  It's no wonder
that it took so many iterations to get here, though.

It looks like most of the docstrings of things in plot_mpl.mpl can be
deleted or shortened.  Delete the parameter lists in mpl._() functions
and link the mainline parameter lists to the official options lists