Fix bug w/ readying pyfun histories when pyfun_hist.dat -> pyfun_hist.00.dat
When you shift from steady-state to time-accurate in FUN3D, it starts over
from iter 0. To avoid loss of history, CAPE renames the old (steady-state)
iterative history file to a different name with the phase number in the
new file name. However, this caused issues with the cached iterative
history because now the name of the file from which the earlier steady-
state iterations came now has a different file name.

As of this commit, when CAPE detects something like this, it just ignores
the .cdb archive