Improve (slightly) logic behind warmstart
There's more work to do here, CaseRunner should probably detect its own
mesh files. This update at least uses CaseRunner for more logic. Also need
a setting to include or not include the *mesh* in the WarmStart (...
alternative is to just copy the solution but not the previous case's mesh)